% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get timeseries data in a tidy dataframe}
\usage{
get_data(subdomain = c("kyy", "ypaat", "emy"), timeID)
}
\arguments{
\item{timeID}{}
}
\description{
Get timeseries data in a tidy dataframe
}
