% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables.R
\name{get_functions}
\alias{get_functions}
\alias{get_stations}
\alias{get_timeseries}
\alias{get_instruments}
\alias{get_water_basins}
\alias{get_water_divisions}
\alias{get_political_divisions}
\alias{get_variables}
\alias{get_units_of_measurement}
\alias{get_time_steps}
\alias{get_owners}
\alias{get_instruments_type}
\alias{get_station_type}
\alias{get_database}
\title{Get tibbles from Hydroscope}
\usage{
get_stations(subdomain = c("kyy", "ypaat", "emy", "deh"), translit = TRUE)

get_timeseries(subdomain = c("kyy", "ypaat", "emy", "deh"), translit = TRUE)

get_instruments(subdomain = c("kyy", "ypaat", "emy", "deh"),
  translit = TRUE)

get_water_basins(subdomain = c("kyy", "ypaat", "emy", "deh"),
  translit = TRUE)

get_water_divisions(subdomain = c("kyy", "ypaat", "emy", "deh"),
  translit = TRUE)

get_political_divisions(subdomain = c("kyy", "ypaat", "emy", "deh"),
  translit = TRUE)

get_variables(subdomain = c("kyy", "ypaat", "emy", "deh"), translit = TRUE)

get_units_of_measurement(subdomain = c("kyy", "ypaat", "emy", "deh"),
  translit = TRUE)

get_time_steps(subdomain = c("kyy", "ypaat", "emy", "deh"), translit = TRUE)

get_owners(subdomain = c("kyy", "ypaat", "emy", "deh"), translit = TRUE)

get_instruments_type(subdomain = c("kyy", "ypaat", "emy", "deh"),
  translit = TRUE)

get_station_type(subdomain = c("kyy", "ypaat", "emy", "deh"),
  translit = TRUE)

get_database(subdomain = c("kyy", "ypaat", "emy", "deh"), translit = TRUE)
}
\arguments{
\item{subdomain}{One of the subdomains of Hydroscope in the vector
\code{c("kyy", "ypaat", "emy", "deh")}.}

\item{translit}{Automatically transliterate Greek to Latin.}
}
\value{
If \code{subdomain} is one of:
\itemize{
\item{\code{kyy}, Ministry of Environment and Energy.}
\item{\code{ypaat}, Ministry of Rural Development and Food.}
\item{\code{deh}, Greek Public Power Corporation.}
\item{\code{emy}, National Meteorological Service.}
}
returns a tibble or a named list with tibbles from the corresponding database
of Hydroscope. Otherwise returns an error message.
}
\description{
A family of functions that return a tibble from a specific database
of Hydroscope using the Enhydris API. \code{get_database} returns a named
list of tibbles using all the family functions.
}
\note{
Objects' IDs are not unique among the different Hydroscope databases. For
example, time series' IDs from http://kyy.hydroscope.gr have same values
with time series' from http://ypaat.hydroscope.gr.

The coordinates of points are based on the European Terrestrial Reference
System 1989 (ETRS89).
}
\examples{

\dontrun{

# get data from all the Hydroscope's subdomains
kyy_db <- get_database("kyy")
ypaat_db <- get_database("ypaat")
emy_db <- get_database("emy")
deh_db <- get_database("deh")

}

}
\references{
The data are retrieved from the Hydroscope's databases:
\itemize{
\item Ministry of Environment, Energy and Climate Change,
\url{http://kyy.hydroscope.gr}
\item Ministry of Rural Development and Food,
\url{http://ypaat.hydroscope.gr}
\item National Meteorological Service,
\url{http://emy.hydroscope.gr}
\item{Greek Public Power Corporation},
\url{http://deh.hydroscope.gr}
}

European Terrestrial Reference System 1989 (ETRS),
\url{http://bit.ly/2kJwFuf}

Greek Water Divisions,
\url{http://bit.ly/2kk0tOm}, \url{http://bit.ly/2ltQC8O}

Greek Water Basins,
\url{http://bit.ly/2Dvzo1W}

Tibble, \url{http://tibble.tidyverse.org/}
}
\author{
Konstantinos Vantas, \email{kon.vantas@gmail.com}
}
