% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{greece_borders}
\alias{greece_borders}
\title{Greek borders}
\format{
A tibble with 18,474 rows and 8 variables:
\describe{
  \item{long}{Longitude in decimal degrees, ETRS89}
  \item{lat}{Latitude in decimal degrees, ETRS89}
  \item{order}{order, integer}
  \item{hole}{hole, boolean}
  \item{piece}{piece, integer}
  \item{group}{group, numeric}
}
}
\source{
Konstantinos Vantas
}
\usage{
greece_borders
}
\description{
The borders of Greece are taken from Geoadata.gov.gr.  The
variables are created using the function tidy from the broom package. This
data frame was created for use with the geom_polygon from ggplot2 package.
}
\keyword{datasets}
