% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_coords.R
\name{hydro_coords}
\alias{hydro_coords}
\title{Convert coordinates from Hydroscope's points}
\usage{
hydro_coords(x)
}
\arguments{
\item{x}{a string vector with the points retrieved from Hydroscope}
}
\value{
a data frame with the longitude and latitude values. If the
coordinates of
}
\description{
\code{hydro_coords} returns a data frame with the stations' longitudes and
latitudes using as input the variable \code{point} from \code{get_stations}
 function.
}
\examples{
\dontrun{
# get stations from the Greek Ministry of Environment and Energy
kyy_stations <-get_stations("kyy")
# create a dataframe with the coords of the stations
coords <- get_coords(kyy_stations$point)
}
}
\author{
Konstantinos Vantas, \email{kon.vantas@gmail.com}
}
