% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydroscoper_package.R
\docType{package}
\name{hydroscoper}
\alias{hydroscoper}
\alias{hydroscoper-package}
\alias{hydroscoper-package}
\title{hydroscoper: Interface to Hydroscope}
\description{
R interface to the  Greek National Data Bank for Hydrological and
Meteorological Information http://www.hydroscope.gr/. It
covers Hydroscope's data sources and provides functions: a) to transliterate
and translate Greek terms to English and b) download time series' data into
tidy dataframes.
}
\section{Data Sources}{


The data are retrieved from the Hydroscope's databases:
\itemize{
\item Ministry of Environment, Energy and Climate Change,
\url{http://kyy.hydroscope.gr}.
\item Ministry of Rural Development and Food,
\url{http://ypaat.hydroscope.gr}.
\item National Meteorological Service,
\url{http://emy.hydroscope.gr}.
\item{Greek Public Power Corporation},
\url{http://deh.hydroscope.gr}.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kvantas/hydroscoper}
  \item Report bugs at \url{https://github.com/kvantas/hydroscoper/issues}
}

}
\author{
\strong{Maintainer}: Konstantinos Vantas \email{kon.vantas@gmail.com} (0000-0001-6387-8791)

}
