% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{hydroscoper_defunct}
\alias{hydroscoper_defunct}
\alias{get_coords}
\title{Defunct functions in hydroscoper}
\usage{
get_coords(...)
}
\arguments{
\item{...}{Defunct function's parameters}
}
\description{
These functions are no longer available in \pkg{hydroscoper}.
}
\details{
Defunct functions:

\itemize{
 \item \code{\link{get_coords}}: This function is defunct. Please use
 \code{\link{hydro_coords}} to convert Hydroscope's points raw format to a
 tidy data frame.
}
}
