% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{timeseries}
\alias{timeseries}
\title{timeseries}
\format{A data frame:
\describe{
    \item{TimeSeriesID}{The time series ID from the database}
    \item{Variable}{The time series variable type}
    \item{TimeStep}{The timestep of time series}
    \item{Unit}{The units of the time series}
    \item{Instrument}{The instrument ID}
    \item{StartDate}{The starting date of time series values}
    \item{EndDate}{The ending date of time series values}
    \item{StationID}{The corresponding station's ID}
    \item{Domain}{The corresponding Hydroscope's domain}
}}
\usage{
timeseries
}
\description{
Time series' data from the Greek National Data bank for Hydrological and
Meteorological Information databases: \url{http://kyy.hydroscope.gr/},
\url{http://emy.hydroscope.gr/} and \url{http://ypaat.hydroscope.gr/}
}
\keyword{datasets}
