% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_coords}
\alias{get_coords}
\title{Create coordinates from Hydroscope's points}
\usage{
get_coords(x)
}
\arguments{
\item{x}{a string vector with the points retrieved from Hydroscope}
}
\value{
a data frame with the longitude and latitude values. If the
coordinates of
}
\description{
\code{get_coords} returns a data frame with the stations' longitudes and
latitudes using as input the variable \code{point} from \code{get_stations}
 function.
}
\examples{
\dontrun{
# get stations from the Greek Ministry of Environment and Energy
df <-get_stations("kyy")
# create a dataframe with the coords of the stations
coords <- get_coords(df$point)
}
}
\author{
Konstantinos Vantas, \email{kon.vantas@gmail.com}
}
