% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get time series values in a tibble}
\usage{
get_data(subdomain = c("kyy", "ypaat", "emy", "deh"), time_id)
}
\arguments{
\item{subdomain}{One of the subdomains of hydroscope.gr}

\item{time_id}{A time series ID}
}
\value{
If \code{subdomain} is one of:
\itemize{
\item{\code{kyy}, Ministry of Environment and Energy}
\item{\code{ypaat}, Ministry of Rural Development and Food}
\item{\code{deh}, Greek Public Power Corporation}
\item{\code{emy}, National Meteorological Service}
}
and \code{time_id} exists in that \code{subdomain}, returns a tibble with the time series values. Otherwise returns an error message.

The dataframe columns are:
\describe{
    \item{date}{The time series Dates (POSIXct)}
    \item{value}{The time series values (numeric)}
    \item{comment}{Comments about the values (character)}
}
}
\description{
\code{get_data} returns a tibble from  a Hydroscope's
time-series text file.
}
\note{
Data are not available freely in the sub-domains:  \code{"deh"}
(Greek Public Power Corporation) and
\code{"emy"} (National Meteorological Service).
}
\examples{
\dontrun{
# get time series 912 from the Greek Ministry of Environment and Energy
time_series <- get_data("kyy", 912)
}

}
\references{
Stations' data are retrieved from the Hydroscope's databases:
\itemize{
\item Ministry of Environment, Energy and Climate Change,
\url{http://kyy.hydroscope.gr}
\item Ministry of Rural Development and Food,
\url{http://ypaat.hydroscope.gr}
}

Tibble, \url{http://tibble.tidyverse.org/}
}
\author{
Konstantinos Vantas, \email{kon.vantas@gmail.com}
}
