% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get time series values in a tidy dataframe}
\usage{
get_data(subdomain = c("kyy", "ypaat"), time_id)
}
\arguments{
\item{subdomain}{One of the subdomains of hydroscope.gr}

\item{time_id}{A time series ID}
}
\value{
If \code{subdomain} is one of \code{"kyy"} (Ministry of Environment
and Energy) or \code{"ypaat"} (Ministry of Rural Development and Food),
and timeID is not NULL, returns a tidy dataframe with the time series values
from the corresponding database of hydroscope.gr. Otherwise gives an error
message.

If the time series ID does not exist in the database, or the url from
Hydroscope could not parsed, displays a warning message and returns a
dataframe with NA values.

The dataframe columns are:
\describe{
    \item{Date}{The time series Dates (POSIXct)}
    \item{Value}{The time series values (numeric)}
    \item{Comment}{Comments about the values (character)}
}
}
\description{
\code{get_data} returns a tidy dataframe with the available data from
a time series in a database of Hydroscope.
}
\note{
The subdomains \code{"deh"} (Greek Public Power Corporation) and \code{"emy"}
(National Meteorological Service) are not used, because the data are not
freely available.

The subdomain "main", which also is not used, is a data
aggregator from several databases. It uses different IDs for stations and
time series.
}
\examples{
\dontrun{
# get time series 912 from the Greek Ministry of Environment and Energy
df <-get_data("kyy", 912)
}

}
\references{
Stations' data are retrieved from the Hydroscope's databases:
\itemize{
\item Ministry of Environment, Energy and Climate Change,
\url{http://kyy.hydroscope.gr}
\item Ministry of Rural Development and Food,
\url{http://ypaat.hydroscope.gr}
}
}
\author{
Konstantinos Vantas, \email{kon.vantas@gmail.com}
}
