% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R, R/get_tables.R
\name{get_functions}
\alias{get_functions}
\alias{get_stations}
\alias{get_timeseries}
\alias{get_instruments}
\alias{get_water_basins}
\alias{get_water_divisions}
\alias{get_political_divisions}
\alias{get_variables}
\alias{get_units_of_measurement}
\alias{get_time_steps}
\alias{get_owners}
\alias{get_instruments_type}
\alias{get_station_type}
\title{Get data frames from Hydroscope}
\usage{
get_stations(subdomain = c("kyy", "ypaat", "emy", "deh"), translit = TRUE)

get_timeseries(subdomain = c("kyy", "ypaat", "emy", "deh"), translit = TRUE)

get_instruments(subdomain = c("kyy", "ypaat", "emy", "deh"),
  translit = TRUE)

get_water_basins(subdomain = c("kyy", "ypaat", "emy", "deh"),
  translit = TRUE)

get_water_divisions(subdomain = c("kyy", "ypaat", "emy", "deh"),
  translit = TRUE)

get_political_divisions(subdomain = c("kyy", "ypaat", "emy", "deh"),
  translit = TRUE)

get_variables(subdomain = c("kyy", "ypaat", "emy", "deh"), translit = TRUE)

get_units_of_measurement(subdomain = c("kyy", "ypaat", "emy", "deh"),
  translit = TRUE)

get_time_steps(subdomain = c("kyy", "ypaat", "emy", "deh"), translit = TRUE)

get_owners(subdomain = c("kyy", "ypaat", "emy", "deh"), translit = TRUE)

get_instruments_type(subdomain = c("kyy", "ypaat", "emy", "deh"),
  translit = TRUE)

get_station_type(subdomain = c("kyy", "ypaat", "emy", "deh"),
  translit = TRUE)
}
\arguments{
\item{subdomain}{One of the subdomains of Hydroscope.}

\item{translit}{Automatically transliterate Greek to Latin.}
}
\value{
If \code{subdomain} is one of:
\itemize{
\item{\code{kyy}, Ministry of Environment and Energy}
\item{\code{ypaat}, Ministry of Rural Development and Food}
\item{\code{deh}, Greek Public Power Corporation}
\item{\code{emy}, National Meteorological Service}
}
returns a tidy dataframe with data from the corresponding database
of Hydroscope. Otherwise returns an error message.
}
\description{
A family of functions that return a data frame from a specific database
of Hydroscope using the Enhydris API.
}
\note{
Objects' IDs are not unique among the different Hydroscope databases. For
example, time series' IDs from http://kyy.hydroscope.gr have same values
with time series' from http://ypaat.hydroscope.gr.

The Greek Water Divisions are:
\tabular{ll}{
\strong{Code}  \tab \strong{Name} \cr
GR01 \tab DYTIKE PELOPONNESOS \cr
GR02 \tab BOREIA PELOPONNESOS \cr
GR03 \tab ANATOLIKE PELOPONNESOS \cr
GR04 \tab DYTIKE STEREA ELLADA \cr
GR05 \tab EPEIROS  \cr
GR06 \tab ATTIKE  \cr
GR07 \tab ANATOLIKE STEREA ELLADA \cr
GR08 \tab THESSALIA  \cr
GR09 \tab DYTIKE MAKEDONIA  \cr
GR10 \tab KENTRIKE MAKEDONIA  \cr
GR11 \tab ANATOLIKE MAKEDONIA  \cr
GR12 \tab THRAKE  \cr
GR13 \tab KRETE  \cr
GR14 \tab NESOI AIGAIOU  \cr
}

The coordinates of points in a stations data frame are based on the European
Terrestrial Reference System 1989 (ETRS89)
}
\examples{

\dontrun{

# get data from the Ministry of Environment and Energy
kyy_stations <- get_stations("kyy")
kyy_timser <- get_timeseries("kyy")
kyy_instr <- get_instruments("kyy")
}

}
\references{
The data are retrieved from the Hydroscope's databases:
\itemize{
\item Ministry of Environment, Energy and Climate Change,
\url{http://kyy.hydroscope.gr}
\item Ministry of Rural Development and Food,
\url{http://ypaat.hydroscope.gr}
\item National Meteorological Service,
\url{http://emy.hydroscope.gr}
\item{Greek Public Power Corporation},
\url{http://deh.hydroscope.gr}
}

European Terrestrial Reference System 1989 (ETRS),
\url{http://bit.ly/2kJwFuf}

Greek Water Divisions,
\url{http://bit.ly/2kk0tOm}, \url{http://bit.ly/2ltQC8O}

Greek Water Basin
\url{http://bit.ly/2Dvzo1W}
}
\author{
Konstantinos Vantas, \email{kon.vantas@gmail.com}
}
