% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries.R
\name{get_timeseries}
\alias{get_timeseries}
\title{Get time series data}
\usage{
get_timeseries(subdomain = c("kyy", "ypaat", "emy", "deh"), translit = TRUE,
  translate = TRUE)
}
\arguments{
\item{subdomain}{One of the subdomains of Hydroscope.}

\item{translit}{Automatically transliterate Greek to Latin.}

\item{translate}{Automatically translates Greek terms to Latin.}
}
\value{
If \code{subdomain} is one of:
\itemize{
\item{\code{kyy}, Ministry of Environment and Energy}
\item{\code{ypaat}, Ministry of Rural Development and Food}
\item{\code{deh}, Greek Public Power Corporation}
\item{\code{emy}, National Meteorological Service}
}
returns a tidy dataframe with time series' data from the corresponding
database of Hydroscope. Otherwise returns an error message.

The dataframe columns are:
\describe{
    \item{timeseries_id}{The time series ID from the domain's database}
    \item{last_modified}{The date that the timeseries data were last
    modified}
    \item{instrument}{The instrument's ID}
    \item{station_id}{The station's ID}
    \item{remarks}{Remarks about the time series}
    \item{variable}{The time series variable}
    \item{unit_of_measurement}{The time series unit of measurement}
    \item{time_step}{The time series time step}

}
}
\description{
\code{get_timeseries} returns a dataframe with data from the time series that
exist in a database of Hydroscope.
}
\note{
Time series' IDs are not unique at the different databases records from
the Hydroscope domains.
}
\examples{
\dontrun{

# get time series' data from the Hydroscope's databases
kyy_ts <- get_timeseries("kyy")
ypaat_ts <- get_timeseries("ypaat")
emy_ts <- get_timeseries("emy")
deh_ts <- get_timeseries("deh")
}

}
\references{
Time series' data are retrieved from the Hydroscope's databases:
\itemize{
\item Ministry of Environment, Energy and Climate Change,
\url{http://kyy.hydroscope.gr}
\item Ministry of Rural Development and Food,
\url{http://ypaat.hydroscope.gr}
\item National Meteorological Service,
\url{http://emy.hydroscope.gr}
\item{Greek Public Power Corporation},
\url{http://deh.hydroscope.gr}
}
}
\author{
Konstantinos Vantas, \email{kon.vantas@gmail.com}
}
