% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_timeseries}
\alias{get_timeseries}
\title{Get timeseries corresponding to a station}
\usage{
get_timeseries(subdomain = c("kyy", "ypaat", "emy", "main"), stationID)
}
\arguments{
\item{stationID}{}
}
\description{
Get timeseries corresponding to a station
}
