% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_timeseries}
\alias{get_timeseries}
\title{Get time series data}
\usage{
get_timeseries(subdomain = c("kyy", "ypaat", "emy", "deh"), translit = TRUE)
}
\arguments{
\item{subdomain}{One of the subdomains of Hydroscope.}

\item{translit}{Automatically transliterate Greek to Latin.}
}
\value{
If \code{subdomain} is one of:
\itemize{
\item{\code{kyy}, Ministry of Environment and Energy}
\item{\code{ypaat}, Ministry of Rural Development and Food}
\item{\code{deh}, Greek Public Power Corporation}
\item{\code{emy}, National Meteorological Service}
}
returns a tidy dataframe with stations' data from the corresponding database
of Hydroscope. Otherwise returns an error message.

The dataframe columns are:
\describe{
    \item{station_id}{The station's ID from the domain's database}
    \item{last_modified}{The date that the station's data were last modified}
    \item{name}{The station's name}
    \item{water_basin}{The station's Water Basin}
    \item{water_division}{The station's Water Division ID}
    \item{political_division}{The station's Political Division}
    \item{altitude}{The station's owner}
    \item{Type}{The station's type}

}
}
\description{
\code{get_stations} returns a dataframe with data from the stations that
exist in a database of Hydroscope.
}
\note{
Stations' IDs might not be unique at the different databases records from the
different Hydroscope domains.
}
\examples{

# get time series' data from the Hydroscope's databases
kyy_ts <- get_timeseries(subdomain = "kyy")
ypaat_ts <- get_timeseries(subdomain = "ypaat")
emy_ts <- get_timeseries(subdomain = "emy")
deh_ts <- get_timeseries(subdomain = "deh")

\dontrun{
ts <- get_timeseries(subdomain = "none")
}

}
\references{
Stations' data are retrieved from the Hydroscope's databases:
\itemize{
\item Ministry of Environment, Energy and Climate Change,
\url{http://kyy.hydroscope.gr}
\item Ministry of Rural Development and Food,
\url{http://ypaat.hydroscope.gr}
\item National Meteorological Service,
\url{http://emy.hydroscope.gr}
\item{Greek Public Power Corporation},
\url{http://deh.hydroscope.gr}
}
}
\author{
Konstantinos Vantas, \email{kon.vantas@gmail.com}
}
