% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hydro_translate}
\alias{hydro_translate}
\title{Translate Greek names and terms to English}
\usage{
hydro_translate(x, value = c("owner", "variable", "timestep", "division"))
}
\arguments{
\item{x}{a string vector}

\item{value}{One of the predefined values in
\code{c("owner", "variable", "timestep", "division")}}
}
\value{
If \code{value} is one of:
\itemize{
\item{\code{owner}, translate organisations' names}
\item{\code{variable}, translate hydrometeorological term}
}
returns a character vector with translations of various hydrometeorological
terms or organisations' names
}
\description{
\code{hydro_translate} Greeek to English transaltions of names and terms
}
\examples{
\dontrun{

# get data from the Ministry of Environment and Energy
kyy_owners <- get_owners("kyy")
kyy_vars <- get_variables("kyy")
owners_names <- hydro_translate(kyy_owners$name, "owner")
vars <- hydro_translate(kyy_vars$descr, "variable")
}
}
