% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydroscoper_package.R
\docType{package}
\name{hydroscoper}
\alias{hydroscoper}
\alias{hydroscoper-package}
\alias{hydroscoper-package}
\title{hydroscoper: Interface to Hydroscope}
\description{
R interface to the  Greek National Data Bank for Hydrological and
Meteorological Information http://www.hydroscope.gr/.
}
\details{
\code{hydroscoper} covers Hydroscope's data sources using the
\href{http://bit.ly/2CLlAAA}{Enhydris API} and provides functions to:
\enumerate{
  \item {Transform the available tables and data sets into
       \href{http://bit.ly/2Dg1sHY}{tibbles}.}
  \item{Transliterate the Greek Unicode names to Latin.}
  \item{Translate various Greek terms to English.}
}
}
\section{Enydris API}{


The Enhydris database is implemented in PostgreSQL. Details about the
database can be found \href{http://bit.ly/2D0cZgA}{here}  and about the
Web-service API \href{http://bit.ly/2FlRtBB}{here}.
}

\section{Data Sources}{


The data are retrieved from the Hydroscope's databases:
\itemize{
  \item \href{http://kyy.hydroscope.gr}{Ministry of Environment, Energy and
  Climate Change}.
  \item \href{http://ypaat.hydroscope.gr}{Ministry of Rural Development
  and Food}.
  \item \href{http://emy.hydroscope.gr}{National Meteorological Service}.
  \item  \href{http://deh.hydroscope.gr}{Greek Public Power Corporation}.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kvantas/hydroscoper}
  \item Report bugs at \url{https://github.com/kvantas/hydroscoper/issues}
}

}
\author{
\strong{Maintainer}: Konstantinos Vantas \email{kon.vantas@gmail.com} (0000-0001-6387-8791)

}
