% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stations}
\alias{stations}
\title{stations}
\format{A data frame:
\describe{
\item{ID}{The station ID in Hydroscope}
\item{Name}{Station Name}
\item{WaterDivisionID}{}
\item{WaterBasin}{}
\item{PoliticalDivision}{}
\item{Owner}{}
\item{Type}{}
\item{Long}{}
\item{Lat}{}
\item{Elevation}{}}}
\usage{
stations
}
\description{
Stations' data from the Greek National Databank for Hydrological and
Meteorological Information, Hydroscope.
}
\keyword{datasets}
