% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{timeseries}
\alias{timeseries}
\title{Timeseries}
\format{A data frame:
\describe{
\item{TimeSeriesID}{The timeseries ID in Hydroscope}
\item{Variable}{Station Name}
\item{TimeStep}{}
\item{Unit}{}
\item{Instrument}{}
\item{StartDate}{}
\item{EndDate}{}
\item{StationID}{}}}
\usage{
timeseries
}
\description{
Timeseries' data from the Greek National Databank for Hydrological and
Meteorological Information, Hydroscope.
}
\keyword{datasets}
