% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{timeseries}
\alias{timeseries}
\title{timeseries}
\format{A tibble with 10,804 rows and 9 variables:
\describe{
    \item{time_id}{The time series ID from the database}
    \item{station_id}{The corresponding station's ID}
    \item{variable}{The time series variable type}
    \item{timestep}{The timestep of time series}
    \item{units}{The units of the time series}
    \item{start_date}{The starting date of time series values}
    \item{end_date}{The ending date of time series values}
    \item{subdomain}{The corresponding Hydroscope's database}
}}
\usage{
timeseries
}
\description{
Time series' data from the Greek National Data Bank for
Hydrological and Meteorological Information.
}
\keyword{datasets}
